1. Remove the rkmedia module dependency 2. Code synchronization: support online capture of RAW images to obtain 3A statistics

+++ rkaiq/multiframe_process.cpp
@@ -51,7 +51,7 @@ void ConverToLE(uint16_t* buf, uint32_t len)
 }
 
 // only for even number frame
-void MultiFrameAverage(uint32_t* pIn1_pOut, uint16_t* POut, uint16_t width, uint16_t height, uint8_t frameNumber)
+void MultiFrameAverage(uint32_t* pIn1_pOut, uint16_t* POut, uint16_t width, uint16_t height, uint16_t frameNumber)
 {
     uint16_t n;
     uint16_t roundOffset = 0;
